Adjusted db connection
