+++ .gitignore
@@ -1,2 +1,3 @@
 *.pyc
+*.env
 frontend/node_modules
