+++ app/models/user.py
@@ -1,4 +1,3 @@
-from card.card import Card
 
 class User:
     def __init__(self, user_id, username=''):
