+++ app/requirements.txt
@@ -3,3 +3,4 @@ flask==2.2.2
 gunicorn==20.1.0
 mysql-connector-python
 pyjwt
+python-dotenv
