+++ app/userCard/db.py
@@ -1,5 +1,7 @@
-from card.card import Card
 import mysql.connector
+
+from database.db import conn
+from models.card import Card
 
 # database connection imported from database/db.py
 
@@ -8,36 +10,30 @@ def pickCard(user_id: int) -> Card:
     Picks a card from the database and returns it
     """
     try:
-        with mysql.connector.connect(
-            host="db-container",
-            user="root",
-            password="palegal",
-            database='langapp'
-        ) as db:
-            query = """
-                    SELECT  cards.id, 
-                            cards.hasAudio, 
-                            GROUP_CONCAT(card_concept.concept_id) AS concepts,
-                            cards.front, 
-                            cards.verse, 
-                            cards.sourceLanguage_id,
-                            cards.targetLanguage_id,
-                            cards.creator_id
-                    FROM cards
-                    LEFT JOIN card_concept
-                        ON cards.id = card_concept.card_id
-                    GROUP BY cards.id, 
-                            cards.hasAudio,
-                            cards.front, 
-                            cards.verse, 
-                            cards.sourceLanguage_id,
-                            cards.targetLanguage_id,
-                            cards.creator_id
-                    LIMIT 1
-                """
-            with db.cursor() as cursor:
-                cursor.execute(query)
-                result = cursor.fetchone()
+        query = """
+                SELECT  cards.id, 
+                        cards.hasAudio, 
+                        GROUP_CONCAT(card_concept.concept_id) AS concepts,
+                        cards.front, 
+                        cards.verse, 
+                        cards.sourceLanguage_id,
+                        cards.targetLanguage_id,
+                        cards.creator_id
+                FROM cards
+                LEFT JOIN card_concept
+                    ON cards.id = card_concept.card_id
+                GROUP BY cards.id, 
+                        cards.hasAudio,
+                        cards.front, 
+                        cards.verse, 
+                        cards.sourceLanguage_id,
+                        cards.targetLanguage_id,
+                        cards.creator_id
+                LIMIT 1
+            """
+        with conn.cursor() as cursor:
+            cursor.execute(query)
+            result = cursor.fetchone()
 
         card = Card(card_id=result[0],
                     hasAudio=result[1],
